Story/frontend touchups

+++ server/recommendation-service/src/main/java/com/travelopedia/fun/recommendation_service/controller/ChatController.java
@@ -1,8 +1,10 @@
 package com.travelopedia.fun.recommendation_service.controller;
 
 
+import com.travelopedia.fun.recommendation_service.clients.CustomerServiceProxy;
 import com.travelopedia.fun.recommendation_service.model.Chat;
 import com.travelopedia.fun.recommendation_service.model.ChatMessage;
+import com.travelopedia.fun.recommendation_service.model.ChatMessageRequest;
 import com.travelopedia.fun.recommendation_service.service.ChatService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -17,19 +19,35 @@ public class ChatController {
     @Autowired
     private ChatService chatService;
 
+    @Autowired
+    private CustomerServiceProxy customerServiceProxy;
+
     @GetMapping("/{conversationId}")
-    public ResponseEntity<List<ChatMessage>> getChatMessages(@PathVariable String conversationId) {
+    public ResponseEntity<List<ChatMessage>> getChatMessages(@RequestHeader("Authorization") String authorization, @PathVariable String conversationId) {
+        String email = customerServiceProxy.getExampleEndpoint(authorization);
+        if(email == null) {
+            return ResponseEntity.status(401).build();
+        }
         return ResponseEntity.ok(chatService.getChatMessagesByConversationId(conversationId));
     }
 
-    @PostMapping("/{conversationId}/messages")
-    public ResponseEntity<ChatMessage> addMessage(@PathVariable String conversationId, @RequestBody ChatMessage message) {
-        message.setConversationId(conversationId);
-        return ResponseEntity.ok(chatService.addMessage(message));
+    @PostMapping("/{conversationId}")
+    public ResponseEntity<ChatMessage> addMessage(@PathVariable String conversationId, @RequestBody ChatMessageRequest chatMessageRequest) {
+
+
+        ChatMessage chatMessage = new ChatMessage();
+        chatMessage.setConversationId(conversationId);
+        chatMessage.setContent(chatMessageRequest.getMessage());
+        chatMessage.setRole(chatMessageRequest.getRole());
+        return ResponseEntity.ok(chatService.addMessage(chatMessage));
     }
 
     @DeleteMapping("/{conversationId}")
-    public ResponseEntity<Void> deleteChat(@PathVariable String conversationId) {
+    public ResponseEntity<Void> deleteChat(@RequestHeader("Authorization") String authorization, @PathVariable String conversationId) {
+        String email = customerServiceProxy.getExampleEndpoint(authorization);
+        if(email == null) {
+            return ResponseEntity.status(401).build();
+        }
         chatService.deleteChat(conversationId);
         return ResponseEntity.ok().build();
     }
